Begin refactoring project to use cobra for command line tool like usage, as well as begin restructing into /pkg and /cmd directories to follow Golang conventions.
