+++ exec/handlers/s3.go
@@ -5,7 +5,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
-	"localstack-quickstart/config"
+	"localstack-quickstart/pkg/config"
 )
 
 type S3Handler struct {
